feat: create pairs
this does not run at this point, it represets a stopping point.

I need to finish the unless() block for when to push a SkillBook in the
Creator.pm.  I need to establish a way to determine if a
SkillBook::Standard already exists in the Array, which since it is
effectively a Tupple is not intuitive.