feat: EvAns scoring
I needed more meta data than I originally expected, and the
read-it-in-from-yaml appraoch turned out to be too limiting despite the
cleanness of some aspects of it.

This is *almost* as clean, the constants are all in relatively few
files, and each in _as few_ files as I can get away with.

I'm deriving as many as I can from existing ones with greps and unions.

I also found a number of bugs as I did the monster class that weren't
obvious until I *had* two classes and especially because of the extra
special cases in the monsters.

I made monsters a class instead of a condition.