build: check point
I needed to move all my models to be name spaced as such
I could not think of another way to avoid a name conflict between the model and controller
For AscendingAttributes.  Specialities remains to get its manager.  Most of the templates
still use the lazy load routes from the CollectionBase and ControllerBase packages.
Generals details is updated to use the Manager.
Conflict Groups are still more notional than actual. Pairs are even more so.