build: check point
index pages for pairs, data not validated yet
the manager of managers is partly built
but how it gets triggered needs revisited.
Mojo apparently has a time based trigger
instead of the activity based trigger currently
in place.  Find and use it.
Need to convert the other managers to be managed
the inherited show and index commands actually cause
problems as the managers are now in place. need to
excise them.
carefully, as they are still useful for rendering
markdown.