build: current state
I have the header and footer working

but I need to change the footer. won't in prod
as it requires a local .git dir.
I need to make it a build script from
the just file that writes to a file I read in
it can and should still use the module I wrote
so the effort is still good, I just need a
different helper class.